+++ flexiznam/schema/__init__.py
@@ -28,6 +28,7 @@ from .harp_data import HarpData
 from .scanimage_data import ScanimageData
 from .microscopy_data import MicroscopyData
 from .onix_data import OnixData
+from .bonsai_data import BonsaiData
 from .neuropixel_data import NeuropixelData
 from .sequencing_data import SequencingData
 from .visstim_data import VisStimData
@@ -37,6 +38,7 @@ Dataset.SUBCLASSES["harp"] = HarpData
 Dataset.SUBCLASSES["scanimage"] = ScanimageData
 Dataset.SUBCLASSES["microscopy"] = MicroscopyData
 Dataset.SUBCLASSES["onix"] = OnixData
+Dataset.SUBCLASSES["bonsai"] = BonsaiData
 Dataset.SUBCLASSES["neuropixel"] = NeuropixelData
 Dataset.SUBCLASSES["sequencing"] = SequencingData
 Dataset.SUBCLASSES["visstim"] = VisStimData